requirements.txt
pandas==2.2.3
protobuf==5.28.3
python-dotenv==1.0.1
Requests==2.32.3
scipy==1.14.1
streamlit==1.39.0
torch==2.4.1
transformers==4.43.2
